backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.3.0
pydantic[email]==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
redis==5.0.4
celery==5.4.0
scikit-learn==1.5.0
xgboost==2.0.3
joblib==1.4.2
httpx==0.27.0
stripe==9.9.0
resend==0.8.0
pytest==8.2.2
